+++ src/Codeception/Codecept.php
@@ -7,7 +7,7 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 class Codecept
 {
-    const VERSION = "2.1.2";
+    const VERSION = "2.1.3";
 
     /**
      * @var \Codeception\PHPUnit\Runner
@@ -187,7 +187,10 @@ class Codecept
             if (empty($config)) {
                 continue;
             }
-            $suiteToRun = "{$suite}-{$envList}";
+            $suiteToRun = $suite;
+            if (!empty($envList)) {
+                $suiteToRun .= ' (' . implode(', ', $envArray) . ')';
+            }
             $this->runSuite($config, $suiteToRun, $test);
         }
     }
@@ -198,7 +201,6 @@ class Codecept
         $suiteManager->initialize();
         $suiteManager->loadTests($test);
         $suiteManager->run($this->runner, $this->result, $this->options);
-
         return $this->result;
     }