+++ src/Codeception/Lib/Connector/Guzzle6.php
@@ -1,7 +1,6 @@
 <?php
 namespace Codeception\Lib\Connector;
 
-use Codeception\Exception\ConnectionException;
 use Codeception\Util\Uri;
 use GuzzleHttp\Client as GuzzleClient;
 use GuzzleHttp\Cookie\CookieJar;
@@ -113,30 +112,32 @@ class Guzzle6 extends Client
         }
 
         $status = $response->getStatusCode();
-        $matches = [];
+        if ($status < 300 || $status >= 400) {
+            $matches = [];
 
-        $matchesMeta = preg_match(
-            '/\<meta[^\>]+http-equiv="refresh" content="(\d*)\s*;?\s*url=(.*?)"/i',
-            $body,
-            $matches
-        );
-
-        if (!$matchesMeta && isset($headers['Refresh'])) {
-            // match by header
-            preg_match(
-                '~(\d*);?url=(.*)~',
-                (string) reset($headers['Refresh']),
+            $matchesMeta = preg_match(
+                '/\<meta[^\>]+http-equiv="refresh" content="(\d*)\s*;?\s*url=(.*?)"/i',
+                $body,
                 $matches
             );
-        }
 
-        if ((!empty($matches)) && (empty($matches[1]) || $matches[1] < $this->refreshMaxInterval)) {
-            $uri = new Psr7Uri($this->getAbsoluteUri($matches[2]));
-            $currentUri = new Psr7Uri($this->getHistory()->current()->getUri());
+            if (!$matchesMeta && isset($headers['Refresh'])) {
+                // match by header
+                preg_match(
+                    '~(\d*);?url=(.*)~',
+                    (string) reset($headers['Refresh']),
+                    $matches
+                );
+            }
 
-            if ($uri->withFragment('') != $currentUri->withFragment('')) {
-                $status = 302;
-                $headers['Location'] = $matchesMeta ? htmlspecialchars_decode($uri) : $uri;
+            if ((!empty($matches)) && (empty($matches[1]) || $matches[1] < $this->refreshMaxInterval)) {
+                $uri = new Psr7Uri($this->getAbsoluteUri($matches[2]));
+                $currentUri = new Psr7Uri($this->getHistory()->current()->getUri());
+
+                if ($uri->withFragment('') != $currentUri->withFragment('')) {
+                    $status = 302;
+                    $headers['Location'] = $matchesMeta ? htmlspecialchars_decode($uri) : (string)$uri;
+                }
             }
         }
 
@@ -148,6 +149,11 @@ class Guzzle6 extends Client
         $baseUri = $this->client->getConfig('base_uri');
         if (strpos($uri, '://') === false) {
             if (strpos($uri, '/') === 0) {
+                $baseUriPath = $baseUri->getPath();
+                if (!empty($baseUriPath) && strpos($uri, $baseUriPath) === 0) {
+                    $uri = substr($uri, strlen($baseUriPath));
+                }
+
                 return Uri::appendPath((string)$baseUri, $uri);
             }
             // relative url
@@ -181,9 +187,6 @@ class Guzzle6 extends Client
 
         try {
             $response = $this->client->send($guzzleRequest, $options);
-        } catch (ConnectException $e) {
-            $url = (string) $this->client->getConfig('base_uri');
-            throw new ConnectionException("Couldn't connect to $url. Please check that web server is running");
         } catch (RequestException $e) {
             if (!$e->hasResponse()) {
                 throw $e;
@@ -291,7 +294,7 @@ class Guzzle6 extends Client
                 }
             } else {
                 $files[] = [
-                    'name' => $name, 
+                    'name' => $name,
                     'contents' => fopen($info, 'r')
                 ];
             }
@@ -299,7 +302,7 @@ class Guzzle6 extends Client
 
         return $files;
     }
-    
+
     protected function extractCookies($host)
     {
         $jar = [];