+++ src/Codeception/Module/REST.php
@@ -9,6 +9,7 @@ use Codeception\Lib\InnerBrowser;
 use Codeception\Lib\Interfaces\DependsOnModule;
 use Codeception\Lib\Interfaces\PartedModule;
 use Codeception\Util\JsonArray;
+use Codeception\Util\JsonType;
 use Codeception\Util\XmlStructure;
 use Symfony\Component\BrowserKit\Cookie;
 use Codeception\Util\Soap as XmlUtils;
@@ -617,13 +618,13 @@ EOF;
      * ``` php
      * <?php
      * // match the first `user.id` in json
-     * $firstUser = $I->grabDataFromJsonResponse('$..users[0].id');
-     * $I->sendPUT('/user', array('id' => $firstUser[0], 'name' => 'davert'));
+     * $firstUserId = $I->grabDataFromResponseByJsonPath('$..users[0].id');
+     * $I->sendPUT('/user', array('id' => $firstUserId[0], 'name' => 'davert'));
      * ?>
      * ```
      *
-     * @param $jsonPath
-     * @return array
+     * @param string $jsonPath
+     * @return array Array of matching items
      * @version 2.0.9
      * @throws \Exception
      * @part json
@@ -755,12 +756,118 @@ EOF;
     public function dontSeeResponseContainsJson($json = [])
     {
         $jsonResponseArray = new JsonArray($this->response);
-        \PHPUnit_Framework_Assert::assertFalse(
+        $this->assertFalse(
             $jsonResponseArray->containsArray($json),
             "Response JSON does not contain JSON provided\n"
             . "- <info>" . var_export($json, true) . "</info>\n"
             . "+ " . var_export($jsonResponseArray->toArray(), true)
         );
+    }
+
+    /**
+     * Checks that Json matches provided types.
+     * In case you don't know the actual values of JSON data returned you can match them by type.
+     * Starts check with a root element. If JSON data is array it will check the first element of an array.
+     * You can specify the path in the json which should be checked with JsonPath
+     *
+     * Basic example:
+     *
+     * ```php
+     * <?php
+     * // {'user_id': 1, 'name': 'davert', 'is_active': false}
+     * $I->seeResponseIsJsonType([
+     *      'user_id' => 'integer',
+     *      'name' => 'string|null',
+     *      'is_active' => 'boolean'
+     * ]);
+     *
+     * // narrow down matching with JsonPath:
+     * // {"users": [{ "name": "davert"}, {"id": 1}]}
+     * $I->seeResponseMatchesJsonType(['name' => 'string'], '$.users[0]');
+     * ?>
+     * ```
+     *
+     * In this case you can match that record contains fields with data types you expected.
+     * The list of possible data types:
+     *
+     * * string
+     * * integer
+     * * float
+     * * array (json object is array as well)
+     * * boolean
+     *
+     * You can also use nested data type structures:
+     *
+     * ```php
+     * <?php
+     * // {'user_id': 1, 'name': 'davert', 'company': {'name': 'Codegyre'}}
+     * $I->seeResponseIsJsonType([
+     *      'user_id' => 'integer|string', // multiple types
+     *      'company' => ['name' => 'string']
+     * ]);
+     * ?>
+     * ```
+     *
+     * You can also apply filters to check values. Filter can be applied with `:` char after the type declatation.
+     *
+     * Here is the list of possible filters:
+     *
+     * * `integer:>{val}` - checks that integer is greater than {val} (works with float and string types too).
+     * * `integer:<{val}` - checks that integer is lower than {val} (works with float and string types too).
+     * * `string:url` - checks that value is valid url.
+     * * `string:regex({val})` - checks that string matches a regex provided with {val}
+     *
+     * This is how filters can be used:
+     *
+     * ```php
+     * <?php
+     * // {'user_id': 1, 'email' => 'davert@codeception.com'}
+     * $I->seeResponseIsJsonType([
+     *      'user_id' => 'string:>0:<1000', // multiple filters can be used
+     *      'email' => 'string:regex(~\@~)' // we just check that @ char is included
+     * ]);
+     *
+     * // {'user_id': '1'}
+     * $I->seeResponseIsJsonType([
+     *      'user_id' => 'string:>0', // works with strings as well
+     * }
+     * ?>
+     * ```
+     *
+     * You can also add custom filters y accessing `JsonType::addCustomFilter` method.
+     * See JsonType reference.
+     *
+     * @version 2.1.3
+     * @param array $jsonType
+     * @part json
+     */
+    public function seeResponseMatchesJsonType(array $jsonType, $jsonPath = null)
+    {
+        $jsonArray = new JsonArray($this->response);
+        if ($jsonPath) {
+            $jsonArray = $jsonArray->filterByJsonPath($jsonPath);
+        }
+        $matched = (new JsonType($jsonArray))->matches($jsonType);
+        $this->assertTrue($matched, $matched);
+    }
+
+    /**
+     * Opposite to `seeResponseMatchesJsonType`.
+     *
+     * @part json
+     * @see seeResponseMatchesJsonType
+     * @param $jsonType jsonType structure
+     * @param null $jsonPath optionally set specific path to structure with JsonPath
+     * @version 2.1.3
+     */
+    public function dontSeeResponseMatchesJsonType($jsonType, $jsonPath = null)
+    {
+        $jsonArray = new JsonArray($this->response);
+        if ($jsonPath) {
+            $jsonArray = $jsonArray->filterByJsonPath($jsonPath);
+        }
+        $matched = (new JsonType($jsonArray))->matches($jsonType);
+        $this->assertNotEquals(true, $matched, sprintf("Unexpectedly the response matched the %s data type", var_export($jsonType, true)));
     }
 
     /**
@@ -952,6 +1059,8 @@ EOF;
     }
 
     /**
+     * Deprecated since 2.0.9 and removed since 2.1.0
+     *
      * @param $path
      * @throws ModuleException
      * @deprecated