+++ src/Codeception/Module/Yii2.php
@@ -7,6 +7,7 @@ use Codeception\Configuration;
 use Codeception\TestCase;
 use Codeception\Lib\Interfaces\ActiveRecord;
 use Codeception\Lib\Interfaces\PartedModule;
+use Codeception\Lib\Interfaces\SupportsDomainRouting;
 use Codeception\Lib\Connector\Yii2 as Yii2Connector;
 use yii\db\ActiveRecordInterface;
 use Yii;
@@ -41,7 +42,7 @@ use Yii;
  * Stability: **stable**
  *
  */
-class Yii2 extends Framework implements ActiveRecord, PartedModule
+class Yii2 extends Framework implements ActiveRecord, PartedModule, SupportsDomainRouting
 {
     /**
      * Application config file must be set.
@@ -67,9 +68,12 @@ class Yii2 extends Framework implements ActiveRecord, PartedModule
     {
         $this->client = new Yii2Connector();
         $this->client->configFile = Configuration::projectDir().$this->config['configFile'];
-        $mainConfig = Configuration::config()['config'];
-        if (isset($mainConfig['test_entry_url'])){
-            $this->client->setServerParameter('HTTPS', parse_url($mainConfig['test_entry_url'], PHP_URL_SCHEME)  === 'https');
+        $mainConfig = Configuration::config();
+        if (isset($mainConfig['config']) && isset($mainConfig['config']['test_entry_url'])){
+            $this->client->setServerParameter(
+                'HTTPS',
+                parse_url($mainConfig['config']['test_entry_url'], PHP_URL_SCHEME) === 'https'
+            );
         }
         $this->app = $this->client->startApp();