+++ tests/data/claypit/tests/skipped/NoEnvironmentCept.php
@@ -1,0 +1,5 @@
+<?php
+// @env nothing
+$I = new SkipGuy($scenario);
+$I->wantTo('run something');
+$I->expect('it will work');